+++ p_client.c
@@ -1512,7 +1512,7 @@ void ClientUserinfoChanged (edict_t *ent, char *userinfo)
 
     elo = Info_ValueForKey(userinfo, "elo");
     if (elo[0]) {
-        ent->client->pers.elo_score = atoi(elo);
+        ent->client->pers.elo_score = atof(elo);
     } else {
         G_StuffCmd(ent, "setu elo \"%d\"\n", DEFAULT_ELO_SCORE);
         G_StuffCmd(ent, "seta elo \"%d\"\n", DEFAULT_ELO_SCORE);
